Refactoring auto completion and commands structure
