+++ src/commands/clean/index.js
