+++ src/commands/scan/index.js
