+++ src/commands/update/index.js
@@ -1,58 +1,9 @@
 'use strict'
 
-const ora = require('ora')
-
-const {
-  execPromise,
-  notify,
-  validatePassword,
-  verifyBin,
-  getUserPassword
-} = require('../../helpers')
+const { notify } = require('../../helpers')
 const updateRepositories = require('./repositories')
-
-const updateSystem = async ({ logs }) => {
-  try {
-    verifyBin(['echo', 'sudo', 'apt'])
-
-    const password = await getUserPassword()
-
-    await validatePassword(password)
-
-    const spinner = ora({
-      text: 'Updating system'
-    })
-
-    !logs && spinner.start()
-
-    await execPromise(`echo ${password} | sudo -S apt update`, { silent: !logs })
-    await execPromise(`echo ${password} | sudo -S apt upgrade -y`, { silent: !logs })
-    await execPromise(`echo ${password} | sudo -S apt autoremove -y`, { silent: !logs })
-    await execPromise(`echo ${password} | sudo -S apt clean`, { silent: !logs })
-
-    !logs && spinner.succeed('System updated')
-
-    notify({ message: 'System updated' })
-  } catch (error) {
-    console.error(error.message)
-  }
-}
-
-const updateMe = async ({ logs }) => {
-  verifyBin(['npm'])
-
-  const spinner = ora({
-    text: 'Updating gr-tools'
-  })
-
-  !logs && spinner.start()
-
-  await execPromise('npm install -g gr-tools@latest', { silent: !logs })
-
-  !logs && spinner.succeed('gr-tools updated')
-
-  notify({ message: 'gr-tools updated' })
-}
+const updateSystem = require('./system')
+const updateMe = require('./me')
 
 const update = (software, commandObject) => {
   try {
